backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database & Vector Store
chromadb==0.4.15
sentence-transformers==2.2.2

# AWS Services
boto3==1.28.62
botocore==1.31.62

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.0

# Data Validation
pydantic==2.5.0

# Environment Management
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
numpy==1.24.3